did a huge refacturing work to optimize the sturcutre of the project. split the crm models in separate files.
started to work on the issue #18
