+++ koalixcrm/crm/exceptions.py
@@ -16,12 +16,14 @@ class UserExtensionMissing(Exception):
     def __str__(self):
         return repr(self.value)
 
+
 class UserExtensionPhoneAddressMissing(Exception):
     def __init__(self, value):
         self.value = value
 
     def __str__(self):
         return repr(self.value)
+
 
 class UserExtensionEmailAddressMissing(Exception):
     def __init__(self, value):
@@ -40,6 +42,14 @@ class OpenInterestAccountMissing(Exception):
 
 
 class IncompleteInvoice(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
+
+
+class InvoiceAlreadyRegistered(Exception):
     def __init__(self, value):
         self.value = value
