+++ koalixcrm/crm/views.py
@@ -10,11 +10,11 @@ from django.utils.translation import ugettext as _
 from koalixcrm.crm.models import *
 
 
-def exportPDF(callingModelAdmin, request, whereToCreateFrom, whatToCreate, redirectTo):
+def exportPDF(calling_model_admin, request, whereToCreateFrom, redirectTo):
     """This method exports PDFs provided by different Models in the crm application
 
         Args:
-          callingModelAdmin (ModelAdmin):  The calling ModelAdmin must be provided for error message response.
+          calling_model_admin (ModelAdmin):  The calling ModelAdmin must be provided for error message response.
           request: The request User is to know where to save the error message
           whereToCreateFrom (Model):  The model from which a PDF should be exported
           whatToCreate (str): What document Type that has to be
@@ -27,30 +27,25 @@ def exportPDF(callingModelAdmin, request, whereToCreateFrom, whatToCreate, redir
         Raises:
           raises Http404 exception if anything goes wrong"""
     try:
-        pdf = whereToCreateFrom.createPDF(whatToCreate)
+        pdf = whereToCreateFrom.createPDF()
         response = HttpResponse(FileWrapper(open(pdf, 'rb')), content_type='application/pdf')
         response['Content-Length'] = path.getsize(pdf)
     except (TemplateSetMissing, UserExtensionMissing, CalledProcessError, UserExtensionEmailAddressMissing, UserExtensionPhoneAddressMissing) as e:
         if isinstance(e, UserExtensionMissing):
             response = HttpResponseRedirect(redirectTo)
-            callingModelAdmin.message_user(request, _("User Extension Missing"))
+            calling_model_admin.message_user(request, _("User Extension Missing"))
         elif isinstance(e, UserExtensionEmailAddressMissing):
             response = HttpResponseRedirect(redirectTo)
-            callingModelAdmin.message_user(request, _("User Extension Email Missing"))
+            calling_model_admin.message_user(request, _("User Extension Email Missing"))
         elif isinstance(e, UserExtensionPhoneAddressMissing):
             response = HttpResponseRedirect(redirectTo)
-            callingModelAdmin.message_user(request, _("User Extension Phone Missing"))
+            calling_model_admin.message_user(request, _("User Extension Phone Missing"))
         elif isinstance(e, TemplateSetMissing):
             response = HttpResponseRedirect(redirectTo)
-            callingModelAdmin.message_user(request, _("Templateset Missing"))
+            calling_model_admin.message_user(request, _("Templateset Missing"))
         elif type(e) == CalledProcessError:
             response = HttpResponseRedirect(redirectTo)
-            callingModelAdmin.message_user(request, e.output)
+            calling_model_admin.message_user(request, e.output)
         else:
             raise Http404
     return response
-
-
-def selectaddress(invoiceid):
-    invoice = Invoice.objects.get(id=invoiceid)
-    address = invoice.contract
